requirements.txt
streamlit
opencv-python
ultralytics
Pillow
scikit-learn
numpy